[ty] Introduce `types::bool`, `types::context_manager` and `types::iteration`
